+++ includes/class-newspack-newsletters.php
@@ -14,8 +14,10 @@ use DrewM\MailChimp\MailChimp;
  */
 final class Newspack_Newsletters {
 
-	const NEWSPACK_NEWSLETTERS_CPT = 'newspack_nl_cpt';
-	const EMAIL_HTML_META          = 'newspack_email_html';
+	const NEWSPACK_NEWSLETTERS_CPT          = 'newspack_nl_cpt';
+	const EMAIL_HTML_META                   = 'newspack_email_html';
+	const NEWSPACK_NEWSLETTERS_PALETTE_META = 'newspack_newsletters_color_palette';
+	const PUBLIC_POST_ID_META               = 'newspack_nl_public_post_id';
 
 	/**
 	 * Supported fonts.
@@ -141,6 +143,14 @@ final class Newspack_Newsletters {
 	 * These have to be registered so the updates are handles correctly.
 	 */
 	public static function register_editor_only_meta() {
+		$default_register_meta_args = [
+			'show_in_rest' => [
+				'schema' => [
+					'context' => [ 'edit' ],
+				],
+			],
+			'type'         => 'string',
+		];
 		$fields = [
 			[
 				'name'               => 'newsletterData',
@@ -158,25 +168,15 @@ final class Newspack_Newsletters {
 			],
 			[
 				'name'               => 'senderName',
-				'register_meta_args' => [
-					'show_in_rest' => [
-						'schema' => [
-							'context' => [ 'edit' ],
-						],
-					],
-					'type'         => 'string',
-				],
+				'register_meta_args' => $default_register_meta_args,
 			],
 			[
 				'name'               => 'senderEmail',
-				'register_meta_args' => [
-					'show_in_rest' => [
-						'schema' => [
-							'context' => [ 'edit' ],
-						],
-					],
-					'type'         => 'string',
-				],
+				'register_meta_args' => $default_register_meta_args,
+			],
+			[
+				'name'               => 'stringifiedLayoutDefaults',
+				'register_meta_args' => $default_register_meta_args,
 			],
 			[
 				'name'               => 'newsletter_send_errors',
@@ -366,8 +366,7 @@ final class Newspack_Newsletters {
 	}
 
 	/**
-	 * Set default layout.
-	 * This can be removed once WP 5.5 adoption is sufficient.
+	 * Set post meta on post creation/save.
 	 *
 	 * @param string  $post_id Numeric ID of the campaign.
 	 * @param WP_Post $post The complete post object.
@@ -375,7 +374,8 @@ final class Newspack_Newsletters {
 	 */
 	public static function save( $post_id, $post, $update ) {
 		if ( ! $update ) {
-			update_post_meta( $post_id, 'template_id', -1 );
+			update_post_meta( $post_id, 'template_id', -1 ); // Set default layout. This can be removed once WP 5.5 adoption is sufficient.
+			update_post_meta( $post_id, self::PUBLIC_POST_ID_META, wp_generate_password( 20, false, false ) ); // Generate a token that can be used to identify this post publicly.
 		}
 	}
 
@@ -681,15 +681,8 @@ final class Newspack_Newsletters {
 	 * @param WP_REST_Request $request API request object.
 	 */
 	public static function api_set_color_palette( $request ) {
-		update_option(
-			'newspack_newsletters_color_palette',
-			wp_json_encode(
-				array_merge(
-					json_decode( (string) get_option( 'newspack_newsletters_color_palette', '{}' ), true ) ?? [],
-					json_decode( $request->get_body(), true )
-				)
-			)
-		);
+		self::update_color_palette( json_decode( $request->get_body(), true ) );
+
 		return \rest_ensure_response( [] );
 	}
 
@@ -752,6 +745,7 @@ final class Newspack_Newsletters {
 					'font_body'        => get_post_meta( $post->ID, 'font_body', true ),
 					'font_header'      => get_post_meta( $post->ID, 'font_header', true ),
 					'custom_css'       => get_post_meta( $post->ID, 'custom_css', true ),
+					'layout_defaults'  => get_post_meta( $post->ID, 'layout_defaults', true ),
 				];
 				return $post;
 			},
@@ -1221,6 +1215,25 @@ final class Newspack_Newsletters {
 				exit;
 			}
 		}
+	}
+
+	/**
+	 * Updates the default newsletters color palette option.
+	 *
+	 * @param array $palette The updated color palette.
+	 *
+	 * @return bool True if the option was updated, false otherwise.
+	 */
+	public static function update_color_palette( $palette ) {
+		return update_option(
+			self::NEWSPACK_NEWSLETTERS_PALETTE_META,
+			wp_json_encode(
+				array_merge(
+					json_decode( (string) get_option( self::NEWSPACK_NEWSLETTERS_PALETTE_META, '{}' ), true ) ?? [],
+					$palette
+				)
+			)
+		);
 	}
 }
 Newspack_Newsletters::instance();