+++ includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp.php
@@ -8,6 +8,7 @@
 defined( 'ABSPATH' ) || exit;
 
 use DrewM\MailChimp\MailChimp;
+use Newspack\Newsletters\Subscription_List;
 use Newspack\Newsletters\Subscription_Lists;
 
 /**
@@ -418,21 +419,25 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 				$mc->get( "campaigns/$mc_campaign_id" ),
 				__( 'Error retrieving Mailchimp campaign.', 'newspack_newsletters' )
 			);
-			$folders             = Newspack_Newsletters_Mailchimp_Cached_Data::get_folders();
-			$list_id             = $campaign && isset( $campaign['recipients']['list_id'] ) ? $campaign['recipients']['list_id'] : null;
-			$merge_fields        = $list_id ? Newspack_Newsletters_Mailchimp_Cached_Data::get_merge_fields( $list_id ) : [];
-			$interest_categories = $list_id ? Newspack_Newsletters_Mailchimp_Cached_Data::get_interest_categories( $list_id ) : null;
-			$segments            = $list_id ? Newspack_Newsletters_Mailchimp_Cached_Data::get_segments( $list_id ) : [];
+			$list_id = $campaign && isset( $campaign['recipients']['list_id'] ) ? $campaign['recipients']['list_id'] : null;
 
-			return [
-				'lists'               => $this->get_lists(),
-				'folders'             => $folders,
-				'merge_fields'        => $merge_fields,
+			$lists = $this->get_lists( true );
+			if ( \is_wp_error( $lists ) ) {
+				return $lists;
+			}
+
+			$newsletter_data = [
 				'campaign'            => $campaign,
 				'campaign_id'         => $mc_campaign_id,
-				'interest_categories' => $interest_categories,
-				'segments'            => $segments,
+				'folders'             => Newspack_Newsletters_Mailchimp_Cached_Data::get_folders(),
+				'interest_categories' => $this->get_interest_categories( $list_id ),
+				'lists'               => $lists,
+				'merge_fields'        => $list_id ? Newspack_Newsletters_Mailchimp_Cached_Data::get_merge_fields( $list_id ) : [],
+				'segments'            => $list_id ? Newspack_Newsletters_Mailchimp_Cached_Data::get_segments( $list_id ) : [],
+				'tags'                => $this->get_tags( $list_id ),
 			];
+
+			return $newsletter_data;
 		} catch ( Exception $e ) {
 			return new WP_Error(
 				'newspack_newsletters_mailchimp_error',
@@ -442,9 +447,14 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	}
 
 	/**
-	 * Get lists.
+	 * Get audiences, groups, and tags that can be configured as subscription lists.
+	 * Reconcile edited names for locally-configured lists.
+	 *
+	 * @param boolean $audiences_only Whether to include groups and tags. If true, only return audiences.
+	 *
+	 * @return array|WP_Error List of subscription lists or error.
 	 */
-	public function get_lists() {
+	public function get_lists( $audiences_only = false ) {
 		try {
 			$mc             = new Mailchimp( $this->api_key() );
 			$lists_response = $this->validate(
@@ -471,15 +481,21 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 					$error_message
 				);
 			}
+
+			if ( $audiences_only ) {
+				return $lists_response['lists'];
+			}
+
 			$lists = [];
 
-			// In addition to Audiences, we also automatically fetch all groups and offer them as Subscription Lists.
+			// In addition to Audiences, we also automatically fetch all groups and tags and offer them as Subscription Lists.
 			// Build the final list inside the loop so groups are added after the list they belong to and we can then represent the hierarchy in the UI.
 			foreach ( $lists_response['lists'] as $list ) {
 
 				$lists[]        = $list;
 				$all_categories = Newspack_Newsletters_Mailchimp_Cached_Data::get_interest_categories( $list['id'] );
 				$all_categories = $all_categories['categories'] ?? [];
+				$all_tags       = Newspack_Newsletters_Mailchimp_Cached_Data::get_tags( $list['id'] ) ?? [];
 
 				foreach ( $all_categories as $found_category ) {
 
@@ -492,7 +508,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 
 					$groups = array_map(
 						function ( $group ) use ( $list ) {
-							$group['id']   = $this->create_group_list_id( $group['id'], $list['id'] );
+							$group['id']   = $this->create_group_or_tag_list_id( $group['id'], $list['id'] );
 							$group['type'] = 'mailchimp-group';
 							return $group;
 						},
@@ -500,7 +516,24 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 					);
 					$lists  = array_merge( $lists, $groups );
 				}
+
+				foreach ( $all_tags as $tag ) {
+					$tag['id']   = $this->create_group_or_tag_list_id( $tag['id'], $list['id'], 'tag' );
+					$tag['type'] = 'mailchimp-tag';
+					$lists[]     = $tag;
+				}
 			}
+
+			// Reconcile edited names for locally-configured lists.
+			$configured_lists = Newspack_Newsletters_Subscription::get_lists_config();
+			if ( ! empty( $configured_lists ) ) {
+				foreach ( $lists as &$list ) {
+					if ( ! empty( $configured_lists[ $list['id'] ]['name'] ) ) {
+						$list['local_name'] = $configured_lists[ $list['id'] ]['name'];
+					}
+				}
+			}
+
 			return $lists;
 		} catch ( Exception $e ) {
 			return new WP_Error(
@@ -508,6 +541,71 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 				$e->getMessage()
 			);
 		}
+	}
+
+	/**
+	 * Get interest categories and their groups.
+	 * Reconcile edited names for locally-configured lists.
+	 *
+	 * @param string $list_id List ID.
+	 *
+	 * @return array
+	 */
+	public function get_interest_categories( $list_id = null ) {
+		if ( ! $list_id ) {
+			return [];
+		}
+		$categories = Newspack_Newsletters_Mailchimp_Cached_Data::get_interest_categories( $list_id );
+		if ( empty( $categories['categories'] ) ) {
+			return [];
+		}
+
+		// Reconcile edited names for locally-configured lists.
+		$configured_lists = Newspack_Newsletters_Subscription::get_lists_config();
+		if ( ! empty( $configured_lists ) ) {
+			foreach ( $categories['categories'] as &$category ) {
+				if ( ! empty( $category['interests']['interests'] ) ) {
+					foreach ( $category['interests']['interests'] as &$interest ) {
+						$local_id = $this->create_group_or_tag_list_id( $interest['id'], $list_id );
+						if ( isset( $configured_lists[ $local_id ]['name'] ) ) {
+							$interest['local_name'] = $configured_lists[ $local_id ]['name'];
+						}
+					}
+				}
+			}
+		}
+
+		return $categories;
+	}
+
+	/**
+	 * Get tags. Reconcile edited names for locally-configured lists.
+	 *
+	 * @param string $list_id List ID.
+	 *
+	 * @return array
+	 */
+	public function get_tags( $list_id = null ) {
+		if ( ! $list_id ) {
+			return [];
+		}
+		$tags = Newspack_Newsletters_Mailchimp_Cached_Data::get_tags( $list_id );
+		if ( empty( $tags ) ) {
+			return [];
+		}
+
+		// Reconcile edited names for locally-configured lists.
+		$configured_lists = Newspack_Newsletters_Subscription::get_lists_config();
+		if ( ! empty( $configured_lists ) ) {
+			foreach ( $tags as &$tag ) {
+				$local_id = $this->create_group_or_tag_list_id( $tag['id'], $list_id, 'tag' );
+				if ( isset( $configured_lists[ $local_id ]['name'] ) ) {
+					$tag['local_name'] = $configured_lists[ $local_id ]['name'];
+				}
+			}
+		}
+
+		return $tags;
 	}
 
 	/**
@@ -1048,6 +1146,62 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	}
 
 	/**
+	 * Add contact to a list with multiple groups and/or tags.
+	 *
+	 * @param array $contact The contact, as for the add_contact method.
+	 * @param array $lists List IDs to add the contact to.
+	 */
+	public function add_contact_with_groups_and_tags( $contact, $lists ) {
+		$results = [];
+		$by_list = [];
+		foreach ( $lists as $list_id ) {
+			$group_or_tag_list = $this->maybe_extract_group_or_tag_list( $list_id );
+			if ( $group_or_tag_list && isset( $group_or_tag_list['type'] ) ) {
+				$list_id = $group_or_tag_list['list_id'];
+				if ( 'group' === $group_or_tag_list['type'] && ! isset( $contact['interests'] ) ) {
+					$contact['interests'] = [];
+				}
+				if ( isset( $by_list[ $list_id ] ) ) {
+					$by_list[ $list_id ][] = $group_or_tag_list;
+				} else {
+					$by_list[ $list_id ] = [ $group_or_tag_list ];
+				}
+			} else {
+				// It might be a local list – the list id has to be extracted from the DB.
+				$local_list = Subscription_List::from_form_id( $list_id );
+				if ( $local_list && $local_list->is_configured_for_provider( $this->service ) ) {
+					$list_settings = $local_list->get_provider_settings( $this->service );
+					if ( $list_settings !== null ) {
+						$list_id = $list_settings['list'];
+						$sublist = [
+							'id'      => $list_settings['tag_id'],
+							'list_id' => $list_id,
+							'type'    => 'tag',
+						];
+						if ( isset( $by_list[ $list_id ] ) ) {
+							$by_list[ $list_id ][] = $sublist;
+						} else {
+							$by_list[ $list_id ] = [ $sublist ];
+						}
+					}
+				}
+			}
+			if ( ! isset( $by_list[ $list_id ] ) ) {
+				// If this is not a group-or-tag list, nor a local list, treat the list ID as a regular list.
+				$by_list[ $list_id ] = [];
+			}
+		}
+		if ( empty( $by_list ) ) {
+			return new WP_Error( 'No lists found.' );
+		}
+
+		foreach ( $by_list as $list_id => $sublists ) {
+			$results[] = $this->add_contact( $contact, $list_id, $sublists );
+		}
+		return $results;
+	}
+
+	/**
 	 * Add contact to a list.
 	 *
 	 * @param array  $contact      {
@@ -1058,10 +1212,11 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	 *    @type string[] $metadata Contact additional metadata. Optional.
 	 * }
 	 * @param string $list_id      List to add the contact to.
+	 * @param string $sublists     An array of groups and/or tags in the list to add the contact to.
 	 *
 	 * @return array|WP_Error Contact data if it was added, or error otherwise.
 	 */
-	public function add_contact( $contact, $list_id = false ) {
+	public function add_contact( $contact, $list_id = false, $sublists = [] ) {
 		if ( false === $list_id ) {
 			return new WP_Error( 'newspack_newsletters_mailchimp_list_id', __( 'Missing list id.' ) );
 		}
@@ -1069,15 +1224,11 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 
 		// If contact was added in this execution, we can return the previous
 		// result and bail.
-		if ( ! empty( self::$contacts_added[ $list_id . $email_address ] ) ) {
-			return self::$contacts_added[ $list_id . $email_address ];
+		$cache_key = $list_id . $email_address . wp_json_encode( $sublists );
+		if ( ! empty( self::$contacts_added[ $cache_key ] ) ) {
+			return self::$contacts_added[ $cache_key ];
 		}
 
-		$list = $this->maybe_extract_group_list( $list_id );
-		if ( $list ) {
-			$list_id  = $list['list_id'];
-			$group_id = $list['group_id'];
-		}
 		$new_contact_status = 'subscribed';
 		if ( isset( $contact['metadata'] ) && ! empty( $contact['metadata']['status'] ) ) {
 			$new_contact_status = $contact['metadata']['status'];
@@ -1154,15 +1305,46 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 				}
 			}
 
-			if ( ! empty( $group_id ) ) {
-				$update_payload['interests'] = [
-					$group_id => true,
-				];
+			$all_tags = [];
+			// If the sublists contain any tags, fetch all tags.
+			// This is needed because MC API expects tag names in the contact update payload, not ids.
+			if ( array_filter(
+				$sublists,
+				function( $sublist ) {
+					return 'tag' === $sublist['type']; }
+			) ) {
+				$all_tags = array_reduce(
+					$mc->get( "lists/$list_id/tag-search" )['tags'],
+					function( $tags, $tag ) {
+						$tags[ $tag['id'] ] = $tag['name'];
+						return $tags;
+					},
+					[]
+				);
+			}
+
+			// Add groups and tags, if any.
+			foreach ( $sublists as $sublist ) {
+				$sublist_id   = $sublist['id'];
+				$sublist_type = $sublist['type'];
+				if ( 'group' === $sublist_type ) {
+					if ( ! isset( $update_payload['interests'] ) ) {
+						$update_payload['interests'] = [];
+					}
+					$update_payload['interests'][ $sublist['id'] ] = true;
+				} elseif ( 'tag' === $sublist_type ) {
+					if ( ! isset( $update_payload['tags'] ) ) {
+						$update_payload['tags'] = [];
+					}
+					if ( isset( $all_tags[ $sublist_id ] ) ) {
+						$update_payload['tags'][] = $all_tags[ $sublist_id ];
+					}
+				}
 			}
 
 			// If we're subscribing the contact to a newsletter, they should have some status
 			// because 'non-subscriber' status can't receive newsletters.
-			if ( ! empty( $group_id ) || ! empty( $list_id ) ) {
+			if ( ! empty( $list_id ) || ! empty( $sublists ) ) {
 				$update_payload['status_if_new'] = $new_contact_status ?? 'subscribed';
 				$update_payload['status']        = $new_contact_status ?? 'subscribed';
 			}
@@ -1198,7 +1380,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 				$e->getMessage()
 			);
 		}
-		self::$contacts_added[ $list_id . $email_address ] = $result;
+		self::$contacts_added[ $cache_key ] = $result;
 		return $result;
 	}
 
@@ -1253,11 +1435,17 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		foreach ( $contact['interests'] as $list_id => $interests ) {
 			foreach ( $interests as $group_id => $active ) {
 				if ( $active ) {
-					$groups_lists[] = $this->create_group_list_id( $group_id, $list_id );
+					$groups_lists[] = $this->create_group_or_tag_list_id( $group_id, $list_id );
 				}
 			}
 		}
-		return array_merge( $audience_lists, $groups_lists );
+		$tags_lists = [];
+		foreach ( $contact['tags'] as $list_id => $tags ) {
+			foreach ( $tags as $tag ) {
+				$tags_lists[] = $this->create_group_or_tag_list_id( $tag['id'], $list_id, 'tag' );
+			}
+		}
+		return array_merge( $audience_lists, $groups_lists, $tags_lists );
 	}
 
 	/**
@@ -1281,33 +1469,19 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		}
 		$mc = new Mailchimp( $this->api_key() );
 		try {
-			foreach ( $lists_to_add as $list_id ) {
-				$list = $this->maybe_extract_group_list( $list_id );
-				// If this is a group list, check if the contact is already subscribed to the group, so we don't make an unnecessary call.
-				if ( $list ) {
-					if ( ! empty( $contact['interests'][ $list['list_id'] ] ) ) {
-						if ( isset( $contact['interests'][ $list['list_id'] ][ $list['group_id'] ] ) && $contact['interests'][ $list['list_id'] ][ $list['group_id'] ] ) {
-							continue;
-						}
-					}
-					// If the group doesn't exits, the regular add_contact call below will take care of adding it.
-				}
-				if ( ! isset( $contact['lists'][ $list_id ] ) ) {
-					$this->add_contact( [ 'email' => $email ], $list_id );
-				} else {
-					$mc->patch( "lists/$list_id/members/" . $contact['lists'][ $list_id ]['contact_id'], [ 'status' => 'subscribed' ] );
-				}
-			}
 			foreach ( $lists_to_remove as $list_id ) {
-				$list = $this->maybe_extract_group_list( $list_id );
+				$list = $this->maybe_extract_group_or_tag_list( $list_id );
 				if ( $list ) {
-					$this->remove_group_from_contact( $email, $list['group_id'], $list['list_id'] );
+					$this->remove_group_or_tag_from_contact( $email, $list['id'], $list['list_id'], $list['type'] );
 					continue;
 				}
 				if ( ! isset( $contact['lists'][ $list_id ] ) ) {
 					continue;
 				}
 				$mc->patch( "lists/$list_id/members/" . $contact['lists'][ $list_id ]['contact_id'], [ 'status' => 'unsubscribed' ] );
+			}
+			if ( ! empty( $lists_to_add ) ) {
+				$this->add_contact_with_groups_and_tags( [ 'email' => $email ], $lists_to_add );
 			}
 		} catch ( \Exception $e ) {
 			return new \WP_Error(
@@ -1350,9 +1524,13 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 					$data[ $key ] = $contact[ $key ];
 				}
 			}
-			$data['tags'][ $contact['list_id'] ]      = $contact['tags'];
-			$data['interests'][ $contact['list_id'] ] = $contact['interests'];
-			$data['lists'][ $contact['list_id'] ]     = [
+			if ( isset( $contact['tags'] ) ) {
+				$data['tags'][ $contact['list_id'] ] = $contact['tags'];
+			}
+			if ( isset( $contact['interests'] ) ) {
+				$data['interests'][ $contact['list_id'] ] = $contact['interests'];
+			}
+			$data['lists'][ $contact['list_id'] ] = [
 				'id'         => $contact['id'], // md5 hash of email.
 				'contact_id' => $contact['contact_id'],
 				'status'     => $contact['status'],
@@ -1445,6 +1623,9 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		if ( ! empty( $context ) && strpos( $context, 'group-' ) === 0 ) {
 			$labels['list_explanation'] = __( 'Mailchimp Group', 'newspack-newsletters' );
 		}
+		if ( ! empty( $context ) && strpos( $context, 'tag-' ) === 0 ) {
+			$labels['list_explanation'] = __( 'Mailchimp Tag', 'newspack-newsletters' );
+		}
 		return $labels;
 	}
 
@@ -1480,33 +1661,37 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	}
 
 	/**
-	 * Creates a group list ID based on the group ID and the list ID
+	 * Creates a list ID based on the type, the ID and the list ID
 	 *
-	 * In Mailchimp, we offer both Audiences and Groups as Subscription Lists. We modify the groups IDs so we can differentiate them from the Audiences IDs.
+	 * In Mailchimp, we offer both Audiences, Groups, and Tags as Subscription Lists. We modify the group and tag IDs so we can differentiate them from the Audiences IDs.
 	 *
-	 * Also, when working with groups, we need to know the list ID, so we add it to the group ID.
+	 * Also, when working with groups or tags, we need to know the list ID, so we add it to the ID.
 	 *
-	 * @param string $group_id The Group ID.
+	 * @param string $item_id The item ID.
 	 * @param string $list_id The List/Audience ID.
+	 * @param string $type 'group' or 'tag'.
 	 * @return string
 	 */
-	public function create_group_list_id( $group_id, $list_id ) {
-		return 'group-' . $group_id . '-' . $list_id;
+	public function create_group_or_tag_list_id( $item_id, $list_id, $type = 'group' ) {
+		return $type . '-' . $item_id . '-' . $list_id;
 	}
 
 	/**
-	 * Extract the group and list ID from an ID created with create_group_list_id
+	 * Extract the group or tag + audience (list) ID from an ID created with create_group_or_tag_list_id
 	 *
 	 * @param string $list_id The list ID.
-	 * @return array|false Array with the group ID and the list ID or false if the ID is not a group list ID.
+	 * @return array|false Array with the group/tag ID and the list ID or false if the ID is not a group or tag list ID.
 	 */
-	public function maybe_extract_group_list( $list_id ) {
-		$pattern = '/^group-([^-]+)-([^-]+)$/';
+	public function maybe_extract_group_or_tag_list( $list_id ) {
+		$pattern = '/^(group|tag)-([^-]+)-([^-]+)$/';
 		if ( preg_match( $pattern, $list_id, $matches ) ) {
-			return [
-				'group_id' => $matches[1],
-				'list_id'  => $matches[2],
+			$extracted_ids = [
+				'id'      => $matches[2],
+				'list_id' => $matches[3],
+				'type'    => $matches[1],
 			];
+
+			return $extracted_ids;
 		}
 		return false;
 	}