+++ src/service-providers/mailchimp/ProviderSidebar.js
@@ -2,87 +2,78 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { Fragment, useEffect, useState } from '@wordpress/element';
+import { withSelect } from '@wordpress/data';
+import { Fragment, useEffect } from '@wordpress/element';
 import { ExternalLink, SelectControl, Spinner, Notice } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
-import { getListInterestsSettings } from './utils';
+import { getGroupOptions, getSegmentOptions } from './utils';
+import SelectControlWithOptGroup from '../../components/select-control-with-optgroup/';
 
-const SegmentsSelection = ( {
-	onUpdate,
-	inFlight,
-	targetField,
-	chosenTarget,
-	availableInterests,
-	availableSegments,
-} ) => {
-	const [ targetId, setTargetId ] = useState( chosenTarget.toString() || '' );
+const getSubAudienceOptions = newsletterData => {
+	const groups = newsletterData?.interest_categories || [];
+	const segments = newsletterData?.segments || [];
+	const tags = newsletterData?.tags || [];
+	let optGroups = [];
 
-	const [ isInitial, setIsInitial ] = useState( true );
-	useEffect( () => {
-		if ( ! isInitial ) {
-			onUpdate( targetId );
-		}
-		setIsInitial( false );
-	}, [ targetId ] );
-
-	let options = [];
-
-	if ( availableInterests.length > 0 ) {
-		options = options.concat( [
-			{
-				label: __( 'Group', 'newspack-newsletters' ),
-				value: 'groups',
-				disabled: true,
-			},
-			...availableInterests,
-		] );
+	if ( groups?.categories?.length > 0 ) {
+		optGroups = optGroups.concat( getGroupOptions( groups ) );
 	}
 
-	if ( availableSegments.length > 0 ) {
-		options = options.concat( [
-			{
-				label: __( 'Segment or tag', 'newspack-newsletters' ),
-				value: 'segments',
-				disabled: true,
-			},
-			...availableSegments.map( segment => ( {
-				label: ` - ${ segment.name }`,
-				value: segment.id.toString(),
-			} ) ),
-		] );
+	if ( tags.length > 0 ) {
+		optGroups.push( {
+			label: __( 'Tags', 'newspack-newsletters' ),
+			options: getSegmentOptions( tags ),
+		} );
 	}
 
-	useEffect( () => {
-		if ( targetId !== '' && ! options.find( option => option.value === targetId ) ) {
-			const foundOption = options.find(
-				option => option.value && option.value === `${ targetField || '' }:${ targetId }`
-			);
-			if ( foundOption ) setTargetId( foundOption.value );
-		}
-	}, [ targetId ] );
+	if ( segments.length > 0 ) {
+		optGroups.push( {
+			label: __( 'Segments', 'newspack-newsletters' ),
+			options: getSegmentOptions( segments ),
+		} );
+	}
+	return optGroups;
+};
 
+const SegmentsSelection = ( { onUpdate, inFlight, value, newsletterData } ) => {
+	const optGroups = getSubAudienceOptions( newsletterData );
+	if ( ! optGroups.length ) {
+		return null;
+	}
 	return (
-		<Fragment>
-			{ options.length ? (
-				<SelectControl
-					label={ __( 'Group, segment, or tag', 'newspack-newsletters' ) }
-					value={ targetId }
-					options={ [
-						{
-							label: __( 'All subscribers in audience', 'newspack-newsletters' ),
-							value: '',
-						},
-						...options,
-					] }
-					onChange={ id => setTargetId( id.toString() ) }
-					disabled={ inFlight }
-				/>
-			) : null }
-		</Fragment>
+		<SelectControlWithOptGroup
+			label={ __( 'Group, Segment, or Tag', 'newspack-newsletters' ) }
+			deselectedOptionLabel={ __( 'All subscribers in audience', 'newspack-newsletters' ) }
+			optgroups={ optGroups }
+			value={ value }
+			onChange={ id => onUpdate( id.toString() ) }
+			disabled={ inFlight }
+		/>
 	);
+};
+
+const getSubAudienceValue = newsletterData => {
+	const recipients = newsletterData.campaign?.recipients;
+
+	const targetIdRawValue =
+		recipients?.segment_opts?.saved_segment_id ||
+		recipients?.segment_opts?.conditions[ 0 ]?.value ||
+		'';
+	const targetId =
+		( Array.isArray( targetIdRawValue ) ? targetIdRawValue[ 0 ] : targetIdRawValue ).toString() ||
+		'';
+	const targetField = recipients?.segment_opts?.conditions?.length
+		? recipients?.segment_opts?.conditions[ 0 ]?.field
+		: '';
+	if ( ! targetField || ! targetId ) {
+		return false;
+	}
+	return 'Interests' === recipients?.segment_opts?.conditions[ 0 ]?.condition_type
+		? `${ targetField || '' }:${ targetId }`
+		: targetId;
 };
 
 const ProviderSidebar = ( {
@@ -92,17 +83,39 @@ const ProviderSidebar = ( {
 	renderPreviewText,
 	inFlight,
 	newsletterData,
+	stringifiedLayoutDefaults,
 	apiFetch,
 	postId,
 	updateMeta,
+	createErrorNotice,
+	meta,
 } ) => {
 	const campaign = newsletterData.campaign;
-	const lists =
-		newsletterData.lists && newsletterData.lists.length
-			? newsletterData.lists.filter( list => list.type !== 'mailchimp-group' )
-			: [];
-	const folders = newsletterData.folders || [];
-	const segments = newsletterData.segments || newsletterData.tags || []; // Keep .tags for backwards compatibility.
+
+	// Separate out audiences from other data item types.
+	const audiences = newsletterData?.lists
+		? newsletterData.lists.filter(
+				list => 'mailchimp-group' !== list.type && 'mailchimp-tag' !== list.type
+		  )
+		: [];
+	const folders = newsletterData?.folders || [];
+
+	useEffect( () => {
+		fetchListsAndSegments();
+	}, [] );
+
+	const fetchListsAndSegments = async () => {
+		try {
+			const response = await apiFetch( {
+				path: `/newspack-newsletters/v1/mailchimp/${ postId }/retrieve`,
+			} );
+			updateMeta( 'newsletterData', response );
+		} catch ( e ) {
+			createErrorNotice(
+				e.message || __( 'Error retrieving campaign information.', 'newspack-newsletters' )
+			);
+		}
+	};
 
 	const setList = listId =>
 		apiFetch( {
@@ -171,6 +184,35 @@ const ProviderSidebar = ( {
 		}
 	}, [ campaign ] );
 
+	// If there is a stringified newsletter data from the layout, use it to set the list and segments.
+	useEffect( () => {
+		try {
+			const layoutDefaults = JSON.parse( stringifiedLayoutDefaults );
+			if ( layoutDefaults.senderEmail && layoutDefaults.senderName ) {
+				const existingSenderData = meta.senderEmail && meta.senderName;
+				if ( ! existingSenderData ) {
+					setSender( {
+						senderName: layoutDefaults.senderName,
+						senderEmail: layoutDefaults.senderEmail,
+					} );
+				}
+			}
+			if ( layoutDefaults.newsletterData?.campaign?.recipients?.list_id ) {
+				const existingListId = newsletterData.campaign?.recipients?.list_id;
+				if ( ! existingListId ) {
+					setList( layoutDefaults.newsletterData?.campaign.recipients.list_id ).then( () => {
+						const subAudienceValue = getSubAudienceValue( layoutDefaults.newsletterData );
+						if ( subAudienceValue ) {
+							updateSegments( subAudienceValue );
+						}
+					} );
+				}
+			}
+		} catch ( e ) {
+			// Ignore it.
+		}
+	}, [ stringifiedLayoutDefaults.length ] );
+
 	if ( ! campaign ) {
 		return (
 			<div className="newspack-newsletters__loading-data">
@@ -190,22 +232,9 @@ const ProviderSidebar = ( {
 		);
 	}
 
-	const { list_id } = campaign.recipients || {};
-	const list = list_id && lists.find( ( { id } ) => list_id === id );
+	const { list_id: audienceId } = campaign.recipients || {};
+	const list = audienceId && audiences.find( ( { id } ) => audienceId === id );
 	const { web_id: listWebId } = list || {};
-
-	const recipients = newsletterData.campaign?.recipients;
-
-	const chosenTarget =
-		recipients?.segment_opts?.saved_segment_id ||
-		recipients?.segment_opts?.conditions[ 0 ]?.value ||
-		'';
-
-	const targetField = recipients?.segment_opts?.conditions?.length
-		? recipients?.segment_opts?.conditions[ 0 ]?.field
-		: '';
-
-	const interestSettings = getListInterestsSettings( newsletterData );
 
 	return (
 		<Fragment>
@@ -233,13 +262,13 @@ const ProviderSidebar = ( {
 			<SelectControl
 				label={ __( 'Audience', 'newspack-newsletters' ) }
 				className="newspack-newsletters__to-selectcontrol"
-				value={ list_id }
+				value={ audienceId }
 				options={ [
 					{
 						value: null,
 						label: __( '-- Select an audience --', 'newspack-newsletters' ),
 					},
-					...lists.map( ( { id, name } ) => ( {
+					...audiences.map( ( { id, name } ) => ( {
 						value: id,
 						label: name,
 					} ) ),
@@ -255,11 +284,8 @@ const ProviderSidebar = ( {
 				</p>
 			) }
 			<SegmentsSelection
-				chosenTarget={ Array.isArray( chosenTarget ) ? chosenTarget[ 0 ] : chosenTarget }
-				targetField={ targetField }
-				availableInterests={ interestSettings.options }
-				availableSegments={ segments.filter( segment => segment.member_count > 0 ) }
-				apiFetch={ apiFetch }
+				value={ getSubAudienceValue( newsletterData ) }
+				newsletterData={ newsletterData }
 				inFlight={ inFlight }
 				onUpdate={ updateSegments }
 			/>
@@ -267,4 +293,11 @@ const ProviderSidebar = ( {
 	);
 };
 
-export default ProviderSidebar;
+const mapStateToProps = select => {
+	const { getEditedPostAttribute } = select( 'core/editor' );
+	return {
+		meta: getEditedPostAttribute( 'meta' ),
+	};
+};
+
+export default withSelect( mapStateToProps )( ProviderSidebar );